refactor: move filters to seperate component
